factor WaitForSingleObject into own file
